export-import import wip

+++ inc/Core/Bootstrap.php
@@ -18,7 +18,8 @@ use Bromate\SecurityApiFirewall\SecurityModules\RestApiRoutes\RoutesTreeReposito
 
 use Bromate\SecurityApiFirewall\Admin\AdminPage;
 use Bromate\SecurityApiFirewall\Admin\Documentation;
-use Bromate\SecurityApiFirewall\Core\Settings\ConfigAjaxController;
+use Bromate\SecurityApiFirewall\Core\Settings\SettingsMigrateAjaxController;
+use Bromate\SecurityApiFirewall\Core\Settings\SettingsRepository;
 use Bromate\SecurityApiFirewall\Logs\LogsAjaxController;
 
 use Bromate\SecurityApiFirewall\Cron\Cron;
@@ -48,12 +49,10 @@ final class Bootstrap {
 
 			SettingsAjaxController::register();
 			RestAuthenticationAjaxController::register();
-
 			IpEntriesAjaxController::register();
 			LogsAjaxController::register();
+			SettingsMigrateAjaxController::register();
 			Documentation::register();
-
-			ConfigAjaxController::register();
 
 		}
 	}
@@ -88,6 +87,9 @@ final class Bootstrap {
 			return;
 		}
 
-		Uninstall::delete_data();
+		if( SettingsRepository::read_option( 'config_delete_data_on_uninstall' ) ) {
+			Uninstall::delete_data();
+		}
+
 	}
 }
